+++ spec/tools_spec.rb
@@ -17,9 +17,9 @@ describe 'hash behaviour' do
   end
 
   it 'allows to pack/unpack' do
-    obj1 = StubClass.new.hash2vars!(@hash)
+    obj1 = StubClass.new.hash2vars(@hash)
     hash = obj1.to_hash
-    obj2 = StubClass.new.hash2vars!(hash)
+    obj2 = StubClass.new.hash2vars(hash)
     obj1.hash.each{|k, v| obj2.hash[k].must_equal(v)}
     obj1.array.must_equal(obj2.array)
     obj1.value.must_equal(obj2.value)